Allow disabling the built-in Go Metrics
The golang Prometheus library exposes a bunch of Go-specific metrics in
the default registry, but these don't really have anything to do with
Clickhouse.

This commit adds an option to disable these default metrics, by way of
creating a new registry instead.

To preserve backwards compatibility, the option is disabled by default
and must be explicit toggled on using the -clickhouse_only flag.